[CC] brisbane_llvm.h brisbane_poly.h brisbane_poly_types.h

+++ apps/saxpy-p/kernel.cl
@@ -1,5 +1,5 @@
 __kernel void saxpy(__global float* restrict Z, float A, __global float* restrict X, __global float* restrict Y) {
-  int id = get_global_id(0);
+  size_t id = get_global_id(0);
   Z[id] = A * X[id] + Y[id];
 }
